backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
redis==5.0.1
requests==2.31.0
python-multipart==0.0.6
tenacity==8.2.3
structlog==24.1.0
psycopg2-binary==2.9.9
alembic==1.12.1
email-validator==2.1.0
black==24.3.0
passlib==1.7.4
bcrypt==3.2.0
aiohttp==3.9.1